Added function to read all objects from a map

+++ Overlord_Map_Visualizer_Project/Overlord_Map_Visualizer/resources/classes/OverlordObject.cs
@@ -21,12 +21,12 @@
     }
     class OverlordObject
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Z { get; set; }
         public OverlordObjectType Type { get; set; }
 
-        public OverlordObject(int x, int y, int z, OverlordObjectType type)
+        public OverlordObject(float x, float y, float z, OverlordObjectType type)
         {
             X = x;
             Y = y;
